requirements.txt
beautifulsoup4==4.12.2
dominate==2.7.0
h5py==3.8.0
imageio==2.27.0
matplotlib==3.7.1
munkres==1.1.4
nd2reader==3.3.0
pandas==2.0.0
Pillow==9.5.0
Requests==2.29.0
scikit_learn==1.2.2
scipy==1.9.1
scikit-image==0.20.0
tqdm==4.66.1
visdom==0.2.4
